refactor(navigation)!: Make contacts screen a list/detail view
The contacts screen has been refactored to use a list/detail pattern. The `ContactsScreen` composable has been replaced with `Contacts` and now manages the display of both the list of conversations and the detail view for a selected conversation using `NavigableListDetailPaneScaffold`.

The `ContactsRoutes.Contacts` route has been removed and `ContactsRoutes.Messages` is now the entry point for the contacts graph. The `MessageScreen` is now displayed in the detail pane.

The `MainAppBar` composable has been renamed to `GlobalAppBar` and the previously private `MainAppBar` component is now public.

Deep link URI for messages has been updated to use query parameters.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>