docs: update bug report template app version instructions
The placeholder text for the "Affected app version" field in the bug report issue template was updated.
Previously, it instructed users to find the version in the `release` section of the repository.
It now directs users to find the app version at the bottom of the `Settings` screen within the app itself.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>