feat(contact)!: add manually verified shared contact support
This commit introduces the ability to import and send manually verified shared contacts.

When importing a shared contact, it is now marked as `manuallyVerified`. The `AddSharedContact` service action has been renamed to `ImportContact` to better reflect this. A new `SendContact` service action has been added to send shared contacts to the radio.

The `NodeEntity` and `Node` models have been updated to include a `manuallyVerified` boolean field, which defaults to false and is only set to true when a contact is scanned or imported manually.

When sending a direct message, if the firmware version is 2.7.12 or newer, the destination contact will be sent as a shared contact instead of favoriting the node. This change aims to improve contact management, especially for devices with newer firmware.

BREAKING CHANGE: The `ServiceAction.AddSharedContact` has been renamed to `ServiceAction.ImportContact`. Consumers of this action will need to update their code. The database schema has been updated to version 21 to include the `manually_verified` column in the `node_entity` table.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>