ci(github-actions): move release environment to correct workflow
The `Release` environment, which may contain secrets or require specific protection rules, was incorrectly applied to the `create-internal-release` and `promote-release` workflows. It is now correctly applied only to the `release` workflow's `prepare-build-info` job, which requires access to this environment.

The environment has been removed from the `create-internal-tag` job in the `create-internal-release.yml` workflow and the `promote` job in the `promote-release.yml` workflow.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>